DB root password + upstart removed for Debian
In order to make sure the DB is going to be created, we need to setup
the mysql root password. So I pass it from project to every class which
need it.
Moreover, I added a condition not to install the mysql-down.conf script
in /etc/init if the OS doesn't have upstart as it returned an error on
Debian Squeeze.